Dodavanje preostalih izmjena

+++ APP/Controllers/RoleController.cs
@@ -7,7 +7,7 @@ using System.Web.Mvc;
 
 namespace APP.Controllers
 {
-    [Authorize(Roles = "Admin")]
+    [Services.Authorize(Roles = "Global administrator")]
     public class RoleController : Controller
     {
         private ApplicationRoleManager _roleManager;
@@ -49,63 +49,115 @@ namespace APP.Controllers
 
         // create a new role
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create(RoleViewModel model)
         {
             if(ModelState.IsValid)
             {
                 var role = new ApplicationRole() { Name = model.Name };
-                await RoleManager.CreateAsync(role);
+                var roleCreated = await RoleManager.CreateAsync(role);
+                if(!roleCreated.Succeeded)
+                {
+                    return RedirectToAction("RoleError", "Error");
+                }
+                return RedirectToAction("Index");
             }
-            
-            return RedirectToAction("Index");   
+
+            return RedirectToAction("RoleError", "Error");
         }
 
        
         public async Task<ActionResult> Edit (string id)
         {
+            if(id == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             var role = await RoleManager.FindByIdAsync(id);
+            if(role == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             return View(new RoleViewModel(role));
         }
 
         // edit a role
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(RoleViewModel model)
         {
-            if(TempData["roleBeforeUpdate"] != null)
+            if(ModelState.IsValid)
             {
-                string roleBeforeUpdate = (string)TempData["roleBeforeUpdate"];
-                if(roleBeforeUpdate != model.Name)
+                if (TempData["roleBeforeUpdate"] != null)
                 {
-                    var role = new ApplicationRole() { Id = model.Id, Name = model.Name };
-                    await RoleManager.UpdateAsync(role);
-                }                
+                    string roleBeforeUpdate = (string)TempData["roleBeforeUpdate"];
+                    if (roleBeforeUpdate != model.Name)
+                    {
+                        var role = new ApplicationRole() { Id = model.Id, Name = model.Name };
+                        var updatedRole = await RoleManager.UpdateAsync(role);
+                        if(!updatedRole.Succeeded)
+                        {
+                            return RedirectToAction("RoleError", "Error");
+                        }
+                    }
+                }
+
+                return RedirectToAction("Index");
             }
-            
-            return RedirectToAction("Index");
+
+            return RedirectToAction("RoleError", "Error");
         }
 
         // see role details
         public async Task<ActionResult> Details(string id)
         {
+            if(id == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             var role = await RoleManager.FindByIdAsync(id);
+            if(role == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             return View(new RoleViewModel(role));
         }
 
 
         public async Task<ActionResult> Delete(string id)
         {
+            if (id == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             var role = await RoleManager.FindByIdAsync(id);
+            if (role == null)
+            {
+                return RedirectToAction("RoleError", "Error");
+            }
             return View(new RoleViewModel(role));
         }
 
         // delete a role
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<ActionResult> Delete(RoleViewModel model)
         {
-            
-            var role = await RoleManager.FindByIdAsync(model.Id); // gets a role by id
-            await RoleManager.DeleteAsync(role);  // deletes the role
-            return RedirectToAction("Index");
+            if(model != null)
+            {
+                var role = await RoleManager.FindByIdAsync(model.Id); // gets a role by id
+                if(role == null)
+                {
+                    return RedirectToAction("RoleError", "Error");
+                }
+                var deletedRole = await RoleManager.DeleteAsync(role);  // deletes the role
+                if(!deletedRole.Succeeded)
+                {
+                    return RedirectToAction("RoleError", "Error");
+                }
+                return RedirectToAction("Index");
+            }
+            return RedirectToAction("RoleError", "Error");
         }
     }
 }
